Export video form logic into hook
