+++ packages/atlas/src/views/studio/VideoWorkspace/VideoWorkspace.tsx
@@ -2,19 +2,33 @@ import { FC, memo, useCallback, useEffect, useState } from 'react'
 
 import { ActionBarProps } from '@/components/ActionBar'
 import { BottomDrawer } from '@/components/_overlays/BottomDrawer'
+import { MintNftFirstTimeModal } from '@/components/_overlays/MintNftFirstTimeModal'
 import { useDisplayDataLostWarning } from '@/hooks/useDisplayDataLostWarning'
 import { useMediaMatch } from '@/hooks/useMediaMatch'
 import { useConfirmationModal } from '@/providers/confirmationModal'
+import { useNftActions } from '@/providers/nftActions/nftActions.hooks'
+import { usePersonalDataStore } from '@/providers/personalData'
 import { VideoWorkspaceFormStatus, useVideoWorkspace, useVideoWorkspaceData } from '@/providers/videoWorkspace'
 
 import { VideoForm } from './VideoForm'
 import { useHandleVideoWorkspaceSubmit } from './VideoWorkspace.hooks'
 
+const MINTING_CONFIRMATION_ID = 'minting-confirmation'
+
 export const VideoWorkspace: FC = memo(() => {
   const [formStatus, setFormStatus] = useState<VideoWorkspaceFormStatus | null>(null)
+  const [showMintConfirmationDialog, setShowMintConfirmationDialog] = useState(false)
+  const [shouldHideMintModal, setShouldHideMintModal] = useState(false)
+  const [txVideoId, setTxVideoId] = useState<string>()
+
+  const mintConfirmationDismissed = usePersonalDataStore((state) =>
+    state.dismissedMessages.some((message) => message.id === MINTING_CONFIRMATION_ID)
+  )
+  const updateMintConfirmationDismiss = usePersonalDataStore((state) => state.actions.updateDismissedMessages)
 
   const { isWorkspaceOpen, setIsWorkspaceOpen, editedVideoInfo } = useVideoWorkspace()
   const { tabData } = useVideoWorkspaceData()
+  const { setNftToMint } = useNftActions()
 
   const [openEditDialog, closeEditDialog] = useConfirmationModal({
     type: 'warning',
@@ -104,17 +118,49 @@ export const VideoWorkspace: FC = memo(() => {
   }
 
   return (
-    <BottomDrawer
-      isOpen={isWorkspaceOpen}
-      onClose={closeVideoWorkspace}
-      title={tabData?.title || 'New video'}
-      titleLabel={editedVideoInfo.isNew || editedVideoInfo.isDraft ? 'New' : 'Edit'}
-      pageTitle={isEdit ? 'Edit video' : 'New video'}
-      actionBar={actionBarProps}
-      fixedScrollbar
-    >
-      <VideoForm setFormStatus={setFormStatus} onSubmit={handleVideoWorkspaceSubmit} />
-    </BottomDrawer>
+    <>
+      <MintNftFirstTimeModal
+        shouldHideNextTime={shouldHideMintModal}
+        onShouldHideNextTime={setShouldHideMintModal}
+        show={showMintConfirmationDialog}
+        onSkip={() => {
+          if (shouldHideMintModal) {
+            updateMintConfirmationDismiss(MINTING_CONFIRMATION_ID, true)
+          }
+          setShowMintConfirmationDialog(false)
+        }}
+        onMint={() => {
+          if (shouldHideMintModal) {
+            updateMintConfirmationDismiss(MINTING_CONFIRMATION_ID, true)
+          }
+          if (txVideoId) {
+            setNftToMint(txVideoId)
+          }
+          setShowMintConfirmationDialog(false)
+        }}
+      />
+      <BottomDrawer
+        isOpen={isWorkspaceOpen}
+        onClose={closeVideoWorkspace}
+        title={tabData?.title || 'New video'}
+        titleLabel={editedVideoInfo.isNew || editedVideoInfo.isDraft ? 'New' : 'Edit'}
+        pageTitle={isEdit ? 'Edit video' : 'New video'}
+        actionBar={actionBarProps}
+        fixedScrollbar
+      >
+        <VideoForm
+          setFormStatus={setFormStatus}
+          onSubmit={(data) =>
+            handleVideoWorkspaceSubmit(data).then((videoId) => {
+              setTxVideoId(videoId)
+              if (!mintConfirmationDismissed && !isEdit && videoId) {
+                setShowMintConfirmationDialog(true)
+              }
+            })
+          }
+        />
+      </BottomDrawer>
+    </>
   )
 })
 VideoWorkspace.displayName = 'VideoWorkspace'
